feat: create UI for goal
