+++ backend/app/api/query.py
@@ -2,6 +2,7 @@ from .transaction.queries import TransactionQuery
 from .category.queries import CategoryQuery
 from .authentication.queries import AuthQuery
 from .dashboard.queries import DashboardQuery
+from .goal.queries import GoalQuery
 import strawberry
 
 
@@ -48,5 +49,13 @@ class Query:
             DashboardQuery: An object that exposes dashboard analytics and statistics queries.
         """
         return DashboardQuery()
+    
+    @strawberry.field(description="Access goal-related queries.")
+    def goal(self) -> GoalQuery:
+        """
+        Returns:
+            GoalQuery: An object that exposes goal-related query fields.
+        """
+        return GoalQuery()
