+++ frontend/components/category/category-form.tsx
@@ -1,13 +1,12 @@
 "use client";
 
-import { useState, useMemo, useEffect } from "react";
+import { useState, useMemo } from "react";
 import { useMutation, useQuery } from "@apollo/client";
 import { motion } from "framer-motion";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as z from "zod";
 import { toast } from "sonner";
-import { ChevronRight, FolderTree } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
 import {
@@ -29,31 +28,19 @@ import {
 } from "@/components/ui/select";
 import { Textarea } from "@/components/ui/textarea";
 import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
-import { Badge } from "@/components/ui/badge";
-import {
   CREATE_CATEGORY_MUTATION,
   UPDATE_CATEGORY_MUTATION,
 } from "@/lib/graphql/category/gql";
 import { GET_CATEGORIES_QUERY } from "@/lib/graphql/category/gql";
 import { CategoryTypeEnum, Category } from "@/types/category";
-
-// Helper type for category with children
-type CategoryWithChildren = Category & { children: CategoryWithChildren[] };
-
-// Constants for category levels
-const MAX_CATEGORY_LEVEL = 3; // Maximum allowed nesting level
+import { CategorySelect } from "./category-select";
 
 // Base form schema without validation
 const baseFormSchema = z.object({
   name: z.string().min(2, "Name must be at least 2 characters"),
   type: z.nativeEnum(CategoryTypeEnum),
   description: z.string().min(2, "Description must be at least 2 characters"),
-  parentId: z.string().optional().nullable(),  
+  parentId: z.string().optional().nullable(),
 });
 
 type FormValues = z.infer<typeof baseFormSchema>;
@@ -68,169 +55,6 @@ interface CategoryFormProps {
   };
   onSuccess?: () => void;
 }
-
-// Helper component to render level indicator
-function LevelIndicator({ level }: { level: number }) {
-  if (level === 0) return null;
-
-  return (
-    <TooltipProvider>
-      <Tooltip>
-        {/* <TooltipTrigger asChild>
-          <span className="text-xs text-muted-foreground ml-1">
-            ({level})
-          </span>
-        </TooltipTrigger> */}
-        <TooltipContent>
-          <p>Subcategory Level {level}</p>
-          <p className="text-xs text-muted-foreground mt-1">
-            {level === MAX_CATEGORY_LEVEL
-              ? "Maximum nesting level reached"
-              : `Can create subcategories up to level ${MAX_CATEGORY_LEVEL}`}
-          </p>
-        </TooltipContent>
-      </Tooltip>
-    </TooltipProvider>
-  );
-}
-
-// Helper function to convert Category to CategoryWithChildren
-function convertToCategoryWithChildren(category: Category): CategoryWithChildren {
-  return {
-    ...category,
-    children: [],
-  };
-}
-
-// Helper function to build category hierarchy
-function buildCategoryHierarchy(
-  categories: Category[]
-): CategoryWithChildren[] {
-  const categoryMap = new Map<string, CategoryWithChildren>();
-  const rootCategories: CategoryWithChildren[] = [];
-
-  // First pass: create map of all categories with empty children arrays
-  categories.forEach((category) => {
-    categoryMap.set(category.id, convertToCategoryWithChildren(category));
-  });
-
-  // Second pass: build hierarchy
-  categories.forEach((category) => {
-    const categoryWithChildren = categoryMap.get(category.id)!;
-    if (category.parentId && categoryMap.has(category.parentId)) {
-      const parent = categoryMap.get(category.parentId)!;
-      parent.children.push(categoryWithChildren);
-    } else {
-      rootCategories.push(categoryWithChildren);
-    }
-  });
-
-  return rootCategories;
-}
-
-// Helper function to calculate category levels with validation
-function calculateCategoryLevels(categories: Category[]): {
-  levels: Map<string, number>;
-  hasInvalidLevels: boolean;
-} {
-  const levels = new Map<string, number>();
-  const categoryMap = new Map<string, Category>();
-  let hasInvalidLevels = false;
-
-  // First, create a map of all categories for easy lookup
-  categories.forEach((category) => {
-    categoryMap.set(category.id, category);
-    levels.set(category.id, 0);
-  });
-
-  // Then, calculate the actual level for each category
-  categories.forEach((category) => {
-    let currentCategory = category;
-    let level = 0;
-    const visited = new Set<string>(); // To detect cycles
-
-    // Traverse up the parent chain until we reach a root category
-    while (
-      currentCategory.parentId &&
-      categoryMap.has(currentCategory.parentId)
-    ) {
-      // Check for cycles
-      if (visited.has(currentCategory.id)) {
-        console.error(
-          `Cycle detected in category hierarchy: ${currentCategory.id}`
-        );
-        hasInvalidLevels = true;
-        break;
-      }
-
-      visited.add(currentCategory.id);
-      level++;
-
-      // Check if level exceeds maximum
-      if (level > MAX_CATEGORY_LEVEL) {
-        console.error(
-          `Category ${currentCategory.id} exceeds maximum level of ${MAX_CATEGORY_LEVEL}`
-        );
-        hasInvalidLevels = true;
-        break;
-      }
-
-      currentCategory = categoryMap.get(currentCategory.parentId)!;
-    }
-
-    levels.set(category.id, level);
-  });
-
-  return { levels, hasInvalidLevels };
-}
-
-// Helper component to render category hierarchy in select
-function CategoryOption({
-  category,
-  level = 0,
-  categoryLevels,
-  onSelect,
-}: {
-  category: CategoryWithChildren;
-  level?: number;
-  categoryLevels: Map<string, number>;
-  onSelect?: (category: CategoryWithChildren) => void;
-}) {
-  const actualLevel = categoryLevels.get(category.id) || 0;
-  const isMaxLevel = actualLevel >= MAX_CATEGORY_LEVEL;
-  const indentSize = 12; // pixels of indentation per level
-
-  return (
-    <>
-      <SelectItem
-        value={category.id}
-        disabled={isMaxLevel}
-        onSelect={() => onSelect?.(category)}
-      >
-        <div className="flex items-center">
-          <div style={{ marginLeft: `${actualLevel * indentSize}px` }}>
-            <span className={isMaxLevel ? "text-muted-foreground" : ""}>
-              {category.name}
-            </span>
-            {/* <LevelIndicator level={actualLevel} /> */}
-          </div>
-        </div>
-      </SelectItem>
-      {category.children.map((child) => (
-        <CategoryOption
-          key={child.id}
-          category={child}
-          level={level + 1}
-          categoryLevels={categoryLevels}
-          onSelect={onSelect}
-        />
-      ))}
-    </>
-  );
-}
-
-// Add a constant for the top-level category value
-const TOP_LEVEL_CATEGORY_VALUE = "TOP_LEVEL";
 
 export function CategoryForm({ initialData, onSuccess }: CategoryFormProps) {
   const [isSubmitting, setIsSubmitting] = useState(false);
@@ -279,27 +103,6 @@ export function CategoryForm({ initialData, onSuccess }: CategoryFormProps) {
           parentId: null,
         },
   });
-
-  // Build category hierarchy
-  const categoryHierarchy = useMemo(
-    () => buildCategoryHierarchy(categories),
-    [categories]
-  );
-
-  // Calculate category levels
-  const { levels: categoryLevels, hasInvalidLevels } = useMemo(
-    () => calculateCategoryLevels(categories),
-    [categories]
-  );
-
-  // Show warning if there are invalid levels
-  useEffect(() => {
-    if (hasInvalidLevels) {
-      toast.warning(
-        `Some categories exceed the maximum nesting level of ${MAX_CATEGORY_LEVEL}. Please reorganize your categories.`
-      );
-    }
-  }, [hasInvalidLevels]);
 
   // Get selected parent category for type validation
   const selectedParentId = form.watch("parentId");
@@ -368,7 +171,11 @@ export function CategoryForm({ initialData, onSuccess }: CategoryFormProps) {
               <FormItem>
                 <FormLabel>Name</FormLabel>
                 <FormControl>
-                  <Input placeholder="Category name" {...field} value={field.value} />
+                  <Input
+                    placeholder="Category name"
+                    {...field}
+                    value={field.value}
+                  />
                 </FormControl>
                 <FormMessage />
               </FormItem>
@@ -382,13 +189,7 @@ export function CategoryForm({ initialData, onSuccess }: CategoryFormProps) {
               <FormItem>
                 <FormLabel>Type</FormLabel>
                 <Select
-                  onValueChange={(value) => {
-                    field.onChange(value);
-                    // Clear parent if type doesn't match
-                    if (selectedParent && selectedParent.type !== value) {
-                      form.setValue("parentId", null);
-                    }
-                  }}
+                  onValueChange={field.onChange}
                   value={field.value}
                   defaultValue={field.value}
                   disabled={!!selectedParent}
@@ -421,97 +222,13 @@ export function CategoryForm({ initialData, onSuccess }: CategoryFormProps) {
             control={form.control}
             name="parentId"
             render={({ field }) => (
-              <FormItem>
-                <div className="flex items-center gap-2">
-                  <FormLabel>Parent Category</FormLabel>
-                  <TooltipProvider>
-                    <Tooltip>
-                      <TooltipTrigger asChild>
-                        <span className="cursor-help">
-                          <FolderTree className="h-4 w-4 text-muted-foreground" />
-                        </span>
-                      </TooltipTrigger>
-                      <TooltipContent>
-                        <p>Select a parent category to create a hierarchy</p>
-                        <p className="text-xs text-muted-foreground mt-1">
-                          Subcategories must match their parent's type
-                        </p>
-                      </TooltipContent>
-                    </Tooltip>
-                  </TooltipProvider>
-                </div>
-                <Select
-                  onValueChange={(value) => {
-                    // Convert TOP_LEVEL_CATEGORY_VALUE to undefined for the form
-                    field.onChange(
-                      value === TOP_LEVEL_CATEGORY_VALUE ? undefined : value
-                    );
-                    // Update type to match parent if selected
-                    if (value && value !== TOP_LEVEL_CATEGORY_VALUE) {
-                      const parent = categories.find(
-                        (cat: Category) => cat.id === value
-                      );
-                      if (parent) {
-                        const parentLevel = categoryLevels.get(parent.id) || 0;
-                        if (parentLevel >= MAX_CATEGORY_LEVEL) {
-                          toast.error(
-                            `Cannot create subcategories at level ${MAX_CATEGORY_LEVEL}`
-                          );
-                          field.onChange(null);
-                          return;
-                        }
-                        form.setValue("type", parent.type);
-                      }
-                    }
-                  }}
-                  value={field.value || TOP_LEVEL_CATEGORY_VALUE}
-                  defaultValue={field.value || TOP_LEVEL_CATEGORY_VALUE}
-                >
-                  <FormControl>
-                    <SelectTrigger>
-                      <SelectValue placeholder="Select a parent category (optional)" />
-                    </SelectTrigger>
-                  </FormControl>
-                  <SelectContent>
-                    <SelectItem value={TOP_LEVEL_CATEGORY_VALUE}>
-                      <div className="flex items-center gap-2">
-                        None
-                        <LevelIndicator level={0} />
-                      </div>
-                    </SelectItem>
-                    {categoryHierarchy.map((category) => (
-                      <CategoryOption
-                        key={category.id}
-                        category={category}
-                        categoryLevels={categoryLevels}
-                        onSelect={(selectedCategory) => {
-                          const level =
-                            categoryLevels.get(selectedCategory.id) || 0;
-                          if (level >= MAX_CATEGORY_LEVEL) {
-                            toast.error(
-                              `Cannot create subcategories at level ${MAX_CATEGORY_LEVEL}`
-                            );
-                          }
-                        }}
-                      />
-                    ))}
-                  </SelectContent>
-                </Select>
-                <FormDescription>
-                  {selectedParent
-                    ? `Creating a subcategory under "${
-                        selectedParent.name
-                      }" (Level ${categoryLevels.get(selectedParent.id) || 0})`
-                    : "Optional: Select a parent category to create a subcategory"}
-                  {hasInvalidLevels && (
-                    <span className="text-destructive block mt-1">
-                      Some categories exceed the maximum nesting level of{" "}
-                      {MAX_CATEGORY_LEVEL}
-                    </span>
-                  )}
-                </FormDescription>
-                <FormMessage />
-              </FormItem>
+              <CategorySelect
+                categories={categories}
+                value={field.value || undefined}
+                onChange={(value) => field.onChange(value || null)}
+                onTypeChange={(type) => form.setValue("type", type)}
+                selectedParent={selectedParent}
+              />
             )}
           />
