+++ frontend/components/goal/goal-table-row.tsx
@@ -1,0 +1,1 @@
+
